+++ examples/gunicorn_thrift/gunicorn_config.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 
 worker_class = "thriftpy_gevent"
-thrift_protocol_factory = "thriftpy.protocol:TBinaryProtocolFactory"
-thrift_transport_factory = "thriftpy.transport:TBufferedTransportFactory"
+thrift_protocol_factory = "thriftpy2.protocol:TBinaryProtocolFactory"
+thrift_transport_factory = "thriftpy2.transport:TBufferedTransportFactory"
 
 errorlog = "-"
 loglevel = "info"